Move `compareSemVers` function to `shared`
